refactor(capture): pinch zoom login 분리하여 useZoom hook 생성
