+++ src/features/capture/hooks/index.ts
@@ -1,1 +1,3 @@
 export { default as useCameraPermission } from "./useCameraPermission";
+export { default as useCamera } from "./useCamera";
+export { default as useZoom } from "./useZoom";
